remove `Base` class
