+++ src/trainer/evaluate.py
@@ -10,28 +10,31 @@ from pathlib import Path
 
 import torch
 from omegaconf import DictConfig
-from torch.utils.data import DataLoader
 
 from MLtools.AnimationMetric.AnimationMetric import AnimationMetric
 from MLtools.CarnivalTools.carnival_tools import CarnivalRTS
 
-from .base import Base
 from .data import get_classes
 
 logger = logging.getLogger(__name__)
 
 
-class Evaluator(Base):
+class Evaluator:
     def __init__(
         self,
         cfg: DictConfig,
         model: torch.nn.Module,
-        dl: DataLoader,
+        dl: torch.utils.data.DataLoader,
+        checkpoint: str,
         rank: int = 0,
     ) -> None:
+
+        logging.debug("Setting up Evaluator")
+
         self.cfg = cfg
         self.model = model
         self.dl = dl
+        self.checkpoint_path = self.cfg.experiment_dir / checkpoint
         self.rank = rank
 
     def _write_temp_rts(
@@ -52,22 +55,37 @@ class Evaluator(Base):
             rts = CarnivalRTS.load_from_dict(dict(zip(classes, pred)))
             rts.save(out_dir / f"{utt_id}.rts")
 
-    def evaluate(self, model: str) -> None:
+    def predict(self) -> None:
+
+        self.model.eval()  # switch off grad engine
+
+        preds = [self.model(x.to(self.rank)).detach() for x, _ in self.dl]
+        utt_ids = [Path(path).stem for path in self.dl.dataset.labels]
+
+        return [(pred, utt_id) for pred, utt_id in zip(preds, utt_ids)]
+
+    def evaluate(self) -> None:
 
         # load model and run inference on test data
-        self.load_checkpoint(self.cfg.experiment_dir / model)
-        predictions, utt_ids = zip(*self.predict(self.dl, test=True))
+        self.load_checkpoint(self.checkpoint_path)
+        predictions, utt_ids = zip(*self.predict())
 
+        # write tensors to file in RTS format
         rts_ref_dir = self.cfg.train.data.y_dir
         temp_out_dir = rts_ref_dir.parent / "eval"
         temp_out_dir.mkdir(exist_ok=True, parents=True)
-
-        # write tensors to file in RTS format
         self._write_temp_rts(predictions, utt_ids, rts_ref_dir, temp_out_dir)
 
+        # log scores (similarity, power, score) to console and file
         am = AnimationMetric(str(rts_ref_dir), str(temp_out_dir))
         am.compute_metric()
         am.log(level="set")
 
         # clean up
         shutil.rmtree(temp_out_dir)
+
+    def load_checkpoint(self, checkpoint_path: Path) -> None:
+
+        logger.debug(f"Loading {checkpoint_path}")
+        checkpoint = torch.load(checkpoint_path)
+        self.model.load_state_dict(checkpoint["model_state"])
